pyproject.toml
[project]
name = "click-data-func"
version = "0.1.0"
description = "Azure Functions app"
requires-python = ">=3.9"
dependencies = [
    "azure-functions",
    "azure-identity",
    "azure-storage-blob",
    "azure-eventhub",
]

[tool.uv]
package = false
